[refactor] : 패키지 분리를 위해 class파일 이동
